requirements.txt
authservice-grpc @ git+https://github.com/alexseipopov/capybaras_controller_authservice_grpc.git@main
blinker==1.7.0
click==8.1.7
flake8==6.1.0
Flask==3.0.0
Flask-Cors==4.0.0
grpcio==1.59.2
grpcio-tools==1.59.2
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
protobuf==4.25.0
pycodestyle==2.11.1
pyflakes==3.1.0
python-dotenv==1.0.0
Werkzeug==3.0.1
git+https://github.com/alexseipopov/user_service_grpc.git@main